requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
facenet-pytorch>=2.5.3
albumentations>=1.3.1
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.65.0
pandas>=2.0.0
scipy>=1.11.0
timm>=0.9.0
streamlit>=1.28.0

